requirements.txt
accelerate==1.1.0
beartype==0.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.24.3
packaging==24.2
pandas==2.2.3
pip==25.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rotary-embedding-torch==0.5.3
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.15.2
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.18.5
wheel==0.45.1
deepspeed==0.16.4
lightning_utilities==0.12.0